Refactor CollapsibleRegion into a slotted/configurable component
